FU-79 refactor: Product 관련 클래스를 Product 패키지 하위로 위치 변경
