+++ src/main/java/com/foru/freebe/product/entity/DiscountType.java
@@ -1,4 +1,4 @@
-package com.foru.freebe.productDiscount.entity;
+package com.foru.freebe.product.entity;
 
 public enum DiscountType {
     RATE,
