+++ src/main/java/com/foru/freebe/product/entity/ProductImage.java
@@ -1,6 +1,4 @@
-package com.foru.freebe.productOption.entity;
-
-import com.foru.freebe.product.entity.Product;
+package com.foru.freebe.product.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,44 +10,36 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProductOption {
+public class ProductImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "product_option_id")
+    @Column(name = "product_image_id")
     private Long id;
 
-    @NotNull
-    private String title;
+    // TODO 추후 원본 이미지 리사이징 로직 구현 후 추가 예정
+    // @NotNull
+    private String thumbnailUrl;
 
     @NotNull
-    private Integer price;
-
-    private String description;
+    private String originUrl;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
     private Product product;
 
-    @Builder
-    private ProductOption(String title, Integer price, String description, Product product) {
-        this.title = title;
-        this.price = price;
-        this.description = description;
+    private ProductImage(String thumbnailUrl, String originUrl, Product product) {
+        this.thumbnailUrl = thumbnailUrl;
+        this.originUrl = originUrl;
         this.product = product;
-
-        validatePrice(price);
     }
 
-    private void validatePrice(Integer price) {
-        if (price <= 0) {
-            throw new IllegalArgumentException("Price must be greater than 0.");
-        }
+    public static ProductImage createProductImage(String thumbnailUrl, String originUrl, Product product) {
+        return new ProductImage(thumbnailUrl, originUrl, product);
     }
 }
