+++ src/main/java/com/foru/freebe/product/respository/ProductComponentRepository.java
@@ -1,8 +1,8 @@
-package com.foru.freebe.productComponent.repository;
+package com.foru.freebe.product.respository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.foru.freebe.productComponent.entity.ProductComponent;
+import com.foru.freebe.product.entity.ProductComponent;
 
 public interface ProductComponentRepository extends JpaRepository<ProductComponent, Long> {
 }
