+++ src/main/java/com/foru/freebe/product/respository/ProductDiscountRepository.java
@@ -1,8 +1,8 @@
-package com.foru.freebe.productDiscount.repository;
+package com.foru.freebe.product.respository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.foru.freebe.productDiscount.entity.ProductDiscount;
+import com.foru.freebe.product.entity.ProductDiscount;
 
 public interface ProductDiscountRepository extends JpaRepository<ProductDiscount, Long> {
 }
