+++ src/main/java/com/foru/freebe/product/respository/ProductOptionRepository.java
@@ -1,8 +1,8 @@
-package com.foru.freebe.productOption.repository;
+package com.foru.freebe.product.respository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.foru.freebe.productOption.entity.ProductOption;
+import com.foru.freebe.product.entity.ProductOption;
 
 public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
 }
