+++ src/main/java/com/foru/freebe/product/service/ProductComponentService.java
@@ -1,9 +1,9 @@
-package com.foru.freebe.productComponent.service;
+package com.foru.freebe.product.service;
 
 import java.util.List;
 
 import com.foru.freebe.product.entity.Product;
-import com.foru.freebe.productComponent.dto.ProductComponentDto;
+import com.foru.freebe.product.dto.ProductComponentDto;
 
 public interface ProductComponentService {
     void registerProductComponent(List<ProductComponentDto> productComponentDtoList, Product product);
