+++ src/main/java/com/foru/freebe/product/service/ProductOptionService.java
@@ -1,9 +1,9 @@
-package com.foru.freebe.productOption.service;
+package com.foru.freebe.product.service;
 
 import java.util.List;
 
 import com.foru.freebe.product.entity.Product;
-import com.foru.freebe.productOption.dto.ProductOptionDto;
+import com.foru.freebe.product.dto.ProductOptionDto;
 
 public interface ProductOptionService {
     void registerProductOption(List<ProductOptionDto> productOptionDtoList, Product product);
